django_project/requirements.txt
prettytable~=3.12.0
Django>=3.0,<4.0
psycopg2-binary>=2.9